requirements.txt
absl-py==1.0.0
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci/argon2-cffi-bindings_1644551690056/work
arrow==1.2.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///C:/ci/beautifulsoup4_1650293025093/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
cachetools==5.0.0
certifi==2022.5.18.1
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer==2.0.9
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dicom2nifti==2.3.0
entrypoints @ file:///C:/ci/entrypoints_1649926621128/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///tmp/build/80754af9/python-fastjsonschema_1620414857593/work/dist
fastremap==1.12.2
fill-voids==2.0.2
fonttools==4.28.4
gdcm==1.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.13.3
importlib-metadata==4.11.2
ipykernel @ file:///C:/ci/ipykernel_1646982785443/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1651601994004/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
itk==5.2.1.post1
itk-core==5.2.1.post1
itk-filtering==5.2.1.post1
itk-io==5.2.1.post1
itk-numerics==5.2.1.post1
itk-registration==5.2.1.post1
itk-segmentation==5.2.1.post1
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema @ file:///C:/ci/jsonschema_1650008058050/work
jupyter @ file:///C:/ci/jupyter_1607685287094/work
jupyter-client @ file:///C:/ci/jupyter_client_1650515989971/work
jupyter-console @ file:///opt/conda/conda-bld/jupyter_console_1647002188872/work
jupyter-core @ file:///C:/ci/jupyter_core_1651656286743/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
matplotlib==3.5.2
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mdai==0.11.0
MedPy==0.4.0
mistune @ file:///C:/ci/mistune_1607359457024/work
nbclient @ file:///C:/ci/nbclient_1650290387259/work
nbconvert @ file:///C:/ci/nbconvert_1649741016669/work
nbformat @ file:///C:/ci/nbformat_1649845125000/work
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929390/work
networkx==2.6.3
nibabel==3.2.1
notebook @ file:///C:/ci/notebook_1651666819623/work
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.4.60
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.2
pydicom==2.2.2
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.1
pyparsing==3.0.6
pyrsistent @ file:///C:/ci/pyrsistent_1636093225342/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets==1.2.0
pywin32==302
pywinpty @ file:///C:/ci_310/pywinpty_1644230983541/work/target/wheels/pywinpty-2.0.2-cp39-none-win_amd64.whl
pyzmq @ file:///C:/ci/pyzmq_1638435148211/work
qtconsole @ file:///opt/conda/conda-bld/qtconsole_1649078897110/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1649073884068/work
requests==2.26.0
requests-oauthlib==1.3.1
retrying==1.3.3
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
SimpleITK==2.1.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado @ file:///C:/ci/terminado_1644322780199/work
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.12.1+cu116
torch-tb-profiler==0.4.0
torchaudio==0.12.1+cu116
torchfile==0.1.0
torchmetrics==0.8.2
torchnet==0.0.4
torchvision==0.13.1
tornado @ file:///C:/ci/tornado_1606924294691/work
tqdm==4.62.3
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.7
visdom==0.1.8.9
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
widgetsnbextension @ file:///C:/ci/widgetsnbextension_1644991377168/work
wincertstore==0.2
zipp==3.7.0
